requirements.txt
napi-py==1.1.4
numpy==1.26.4
opencv-python==4.9.0.80
pylzma==0.5.0
